Address review feedback: split predicate.rs to meet <300 line requirement
Extracted unit tests from predicate.rs (237 lines) into separate
predicate_tests.rs module per Judge review feedback.

Module line counts now:
- selection.rs: 231 lines ✓
- predicate.rs: 293 lines ✓ (was 528)
- execution.rs: 207 lines ✓
- mod.rs: 24 lines ✓
- predicate_tests.rs: 236 lines (new)

All modules now meet the <300 line requirement from issue #1912.
All 950 tests still pass.
